Rename rsync server and wire packages
